Add CommitOffsets method to commit multiple partitions at once
This commit adds a new method to GroupClient: CommitOffsets()
The main purpose of the method is to do a bulk commit of the offsets
by providing a map partition -> offset.

The current way when committing happenning for the only partition is left
without any changes due to backward compatibility.

Because of sending more than one partition in a single request, response
from the Kafka will be changed - for those requests, it will respond
with a list of partitions which were used in the CommitOffset request.