add logic to forget cached resolved srv record on call failure
client caches srv record resolutions. Add logic to clear that cache when
a call to one of the brokers fails (likely brokers were deleted /
moved). This is not a "fatal" flaw but results in needless retries.

Change client.LookupSrv and client.RandomBroker to lookupSrv and
randomBroker (there is no need for them to be public).