Add panic to randomBroker() if provided SRV record can't be resolved
`randomBroker()` accepts a name to be resolved for further client
calls. It could be a single broker name, or an SRV record. If it's
an SRV record that for some reasons resolves into an empty list of
hosts, `randomBroker()` will panic.
If provided name is not an SRV, it will be used as is.